+++ scripts/argmaps_init_script.sh
@@ -27,6 +27,8 @@ export PROJECT_DIR
 
 export PANDOC_DATA_DIR="$CONDA_PREFIX/share/pandoc"
 
+export MAPJS_JSON_INPUT_DIR="output/mapjs-json-input/"
+
 # TODO: Remove WP from variable names?
 export MJS_WP_HOME="$WORKSPACE/mapjs-example"
 # export MJS_WP_HOME="$WORKSPACE/mapjs"
@@ -50,11 +52,12 @@ eval "$(pandoc --bash-completion)"
 # TESTING INIT
 
 # TODO: could move to separate test/init_script_argmap_test file
-export INPUT_FILE_YML=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml
-export INPUT_FILE_MUP=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified.mup
-export INPUT_FILE_MD0=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified_0mapjs.md
-export INPUT_FILE_MD=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified_1mapjs.md
-export INPUT_FILE_MD2=$WORKSPACE/Input/Example1_ClearlyFalse_WhiteSwan_simplified_2mapjs.md
-# export INPUT_FILE_JSON=$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.json
+export INPUT_FILE_YML=$WORKSPACE/test/input/Example1_ClearlyFalse_WhiteSwan_simplified.yml
+export INPUT_FILE_MUP=$WORKSPACE/test/input/Example1_ClearlyFalse_WhiteSwan_simplified.mup
+export INPUT_FILE_MD0=$WORKSPACE/test/input/Example1_ClearlyFalse_WhiteSwan_simplified_0mapjs.md
+export INPUT_FILE_MD=$WORKSPACE/test/input/Example1_ClearlyFalse_WhiteSwan_simplified_1mapjs.md
+export INPUT_FILE_MD2=$WORKSPACE/test/input/Example1_ClearlyFalse_WhiteSwan_simplified_2mapjs.md
+export INPUT_FILE_MD_META=$WORKSPACE/test/input/Example1_ClearlyFalse_WhiteSwan_simplified_meta_mapjs.md
+# export INPUT_FILE_JSON=$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified.json
 
 source "$WORKSPACE/test/bash_aliases_argmap_test.sh"