+++ scripts/bash_aliases_argmap.sh
@@ -9,6 +9,7 @@ echo "Running ${BASH_SOURCE[0]}"
 __chrome-mini() {
   # &suffix runs it in background so terminal not blocked
   google-chrome --no-default-browser-check --window-size=500,720 "$1" 2>/dev/null &
+  disown # stops browser blocking terminal and allows all tabs to open in single window.
 }
 
 # Used by pre-commit hook
@@ -16,8 +17,9 @@ __reset_repo() {
   echo 'Restoring output folder to match remote.'
   # Would like to do this for mapjs-example, but risky that it would affect files with legitimate changes.
   # TODO: move output files to a sub-folder.
-  git checkout -- "$WORKSPACE/Output/"
   git checkout -- "$WORKSPACE/examples/"
+  git checkout -- "$WORKSPACE/Output/"
+  git checkout -- "$WORKSPACE/test/output"
 }
 
 # Checks `src/`` for lua files with leftover test/debug code.
@@ -33,8 +35,16 @@ __save_env() {
 }
 
 __clean_repo() {
-  rm "$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.yml"
-  rm "$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup"
+  rm "$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified.yml"
+  rm "$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified.mup"
+  rm "$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified.json"
+  rm "$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified.tex"
+  rm "$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified_0mapjs.pdf"
+  rm "$MJS_WP_HOME/Example1_ClearlyFalse_WhiteSwan_simplified_0mapjs.html"
+  rm "$MJS_WP_HOME/Example1_ClearlyFalse_WhiteSwan_simplified_1mapjs.html"
+  rm "$MJS_WP_HOME/Example1_ClearlyFalse_WhiteSwan_simplified_2mapjs.html"
+  rm "$MJS_WP_HOME/Example1_ClearlyFalse_WhiteSwan_simplified_meta_mapjs.html"
+  rm "$MJS_WP_HOME/f54eea6ed0c060c9d27e1fe3507bfdd75e3e60d4.png"
   # rm "$INPUT_FILE_JSON"
 }
 
@@ -45,17 +55,17 @@ __pack_mapjs() {
   npm --prefix "$MJS_WP_HOME" run pack-js # -- --env.input_map="$1"
 }
 
-# lua argmap2mup Input/Example1_ClearlyFalse_WhiteSwan_simplified.yml > Output/Example1_ClearlyFalse_WhiteSwan_simplified.mup
+# lua argmap2mup test/input/Example1_ClearlyFalse_WhiteSwan_simplified.yml > test/output/Example1_ClearlyFalse_WhiteSwan_simplified.mup
 # TODO add option for .mup vs .json output
-a2m() {                                    # a2m Output/Example1_simple.yml
+a2m() {                                    # a2m test/output/Example1_simple.yml
   NAME=$(basename --suffix=".yml" "$1") && # && ensures error failure stops remaining commands.
-    OUTPUT=${2:-$WORKSPACE/Output/$NAME.mup} &&
+    OUTPUT=${2:-$WORKSPACE/test/output/$NAME.mup} &&
     argmap2mup "$1" >"$OUTPUT" &&
     # TODO: Should return $2 value so can be used by calling app e.g. a2mo or a2mj.
     echo "Generated: $OUTPUT"
 }
 
-a2mu() { # a2mu Output/Example1_simple.yml
+a2mu() { # a2mu test/output/Example1_simple.yml
   NAME=$(basename --suffix=".yml" "$1") &&
     argmap2mup --upload --name "$NAME.mup" --folder 1cSnE4jv5f1InNVgYg354xRwVPY6CvD0x "$1" &&
     echo "Uploaded: $1 to GDrive."
@@ -63,36 +73,39 @@ a2mu() { # a2mu Output/Example1_simple.yml
 
 a2mo() { # Deprecated, use a2jo instead.
   NAME=$(basename --suffix=".yml" "$1") &&
-    OUTPUT=${2:-$WORKSPACE/Output/$NAME.json} &&
+    OUTPUT=${2:-$WORKSPACE/test/output/$NAME.json} &&
     a2m "$1" "$OUTPUT" &&
     __pack_mapjs "$OUTPUT"
 }
 
-a2jo() { # m2a Output/Example1_simple.yml
+# TODO is this necessary any longer?
+a2jo() { # m2a output/mapjs-json-input/Example1_simple.yml
   NAME=$(basename --suffix=".yml" "$1")
-  OUTPUT=${2:-$WORKSPACE/Output/$NAME.json}
+  OUTPUT=${2:-$WORKSPACE/test/output/$NAME.json}
   #TODO: Is there a way to pipe a2m output directly into pack_mapjs?
   # Then __pack_mapjs should ideally have been generated during install/init, so won't be needed to be called after, will just be referenced in js
   a2m "$1" "$OUTPUT" &&
     __pack_mapjs # "$OUTPUT"
 }
 
-m2a() { # m2a Output/Example1_simple.mup
+m2a() { # m2a test/output/Example1_simple.mup
   NAME=$(basename --suffix=".mup" "$1")
-  OUTPUT=${2:-$WORKSPACE/Output/$NAME.yml}
+  OUTPUT=${2:-$WORKSPACE/test/output/$NAME.yml}
   mup2argmap "$1" >"$OUTPUT" &&
     echo "Generated: $OUTPUT"
 }
 
-a2t() { # a2t Output/Example1_simple.yml
+a2t() { # a2t test/output/Example1_simple.yml
   NAME=$(basename --suffix=".yml" "$1") &&
-    argmap2tikz "$1" >"${2:-$WORKSPACE/Output/$NAME.tex}" &&
-    echo "Generated: ${2:-$WORKSPACE/Output/$NAME.tex}"
+    argmap2tikz "$1" >"${2:-$WORKSPACE/test/output/$NAME.tex}" &&
+    echo "Generated: ${2:-$WORKSPACE/test/output/$NAME.tex}"
 }
 
-md2hf() { # md2h Input/example.md
+md2hf() { # md2h test/input/example.md
   NAME=$(basename --suffix=".md" "$1")
-  OUTPUT=${2:-$MJS_WP_HOME/index.html}
+  # TODO: get js file reference to work from other locations so I can write here:
+  # OUTPUT=${2:-$WORKSPACE/test/output/$NAME.html}
+  OUTPUT=${2:-$MJS_WP_HOME/$NAME.html}
   # QUESTION: Is it worth putting some of these settings into a meta-data or defaults file?
   # If so, how would I easily update it? Or just put it in relevant root of mapjs folder?
   # Needed? --metadata=curdir:"$MJS_WP_HOME"
@@ -107,17 +120,17 @@ md2hf() { # md2h Input/example.md
 
   # TODO: Shouldn't need this part, but will need to run __pack_mapjs after updating start.js.
   # call a2jo?
-  # __pack_mapjs # "$WORKSPACE/Output/Example1_ClearlyFalse_WhiteSwan_simplified.json"
+  # __pack_mapjs # "$WORKSPACE/test/output/Example1_ClearlyFalse_WhiteSwan_simplified.json"
   __chrome-mini "$OUTPUT"
 }
 
 # This is meant to output an html doc fragment rather than full doc, so removing template.
-# TODO: fix, this currently creates html output in Output folder: e.g. file:///home/s6mike/git_projects/argmap/Output/example-updated.html
-# Which breaks links to webpack output js, looks in: file:///home/s6mike/git_projects/argmap/Output/site/main.js
+# TODO: fix, this currently creates html output in test/output folder: e.g. file:///home/s6mike/git_projects/argmap/test/output/example-updated.html
+# Which breaks links to webpack output js, looks in: file:///home/s6mike/git_projects/argmap/test/output/site/main.js
 # Probably because I'm now using a relative link to the js file, so that I can view in main chrome browser.
-md2htm() { # md2htm Input/example-updated.md
+md2htm() { # md2htm test/input/example-updated.md
   NAME=$(basename --suffix=".md" "$1")
-  OUTPUT=${2:-$WORKSPACE/Output/$NAME.html}
+  OUTPUT=${2:-$WORKSPACE/test/output/$NAME.html}
 
   # TODO: Put this into new function?
   # Or use a defaults file:
@@ -125,18 +138,18 @@ md2htm() { # md2htm Input/example-updated.md
   pandoc "$1" -o "$OUTPUT" --lua-filter="$WORKSPACE/src/pandoc-argmap.lua" --data-dir="$PANDOC_DATA_DIR" >/dev/null &&
     echo "Generated: $OUTPUT"
   wait # waits for png to appear
-  mv ./*.png "$WORKSPACE/Output/"
+  mv ./*.png "$WORKSPACE/test/output/"
   __chrome-mini "$OUTPUT"
 }
 
-md2pdf() { # md2pdf Input/example.md
+md2pdf() { # md2pdf test/input/example.md
   NAME=$(basename --suffix=".md" "$1")
-  OUTPUT=${2:-$WORKSPACE/Output/$NAME.pdf}
+  OUTPUT=${2:-$WORKSPACE/test/output/$NAME.pdf}
   pandoc "$1" -o "$OUTPUT" --lua-filter="$WORKSPACE/src/pandoc-argmap.lua" --pdf-engine lualatex --template examples/example-template.latex --data-dir="$PANDOC_DATA_DIR" >/dev/null &&
     echo "Generated: $OUTPUT"
   __chrome-mini "$OUTPUT"
 }
 
 ## Mark functions for export to use in other scripts:
-export -f __reset_repo __clean_repo __check_repo __chrome-mini __save_env __pack_mapjs
+export -f __reset_repo __clean_repo __check_repo __chrome-mini __save_env __pack_mapjs # __chrome-newtab
 export -f a2m m2a a2t a2mu a2mo a2jo md2htm md2hf md2pdf