+++ scripts/install.sh
@@ -158,8 +158,8 @@ cd "$MJS_WP_HOME" || {
   exit 1
 }
 
-# Link src folder to Output folder.
-ln -s "$WORKSPACE/Output" "$MJS_WP_HOME/src/argmap_output"
+# Link src folder to test/output folder.
+ln -s "$WORKSPACE/$MAPJS_JSON_INPUT_DIR" "$MJS_WP_HOME/src/argmap_output"
 
 #nodejs installed with conda